feature: view 세분화
